Updated name from dart_explore to dallasbymetro
Also did some small cleaning
