+++ src/main/java/com/dallasbymetro/backend/entity/StationColor.java
@@ -1,0 +1,5 @@
+package com.dallasbymetro.backend.entity;
+
+public enum StationColor {
+    BLUE, RED, ORANGE, GREEN
+}
