+++ src/main/java/com/dallasbymetro/backend/exception/DartExploreException.java
@@ -1,0 +1,7 @@
+package com.dallasbymetro.backend.exception;
+
+public class DartExploreException extends Exception {
+    public DartExploreException(String message) {
+        super(message);
+    }
+}
