--- src/main/java/com/dart/explore/dto/AmenityDTO.java
@@ -1,43 +1,0 @@
-package com.dart.explore.dto;
-
-import com.dart.explore.entity.Amenity;
-import jakarta.validation.constraints.NotNull;
-
-public class AmenityDTO {
-    @NotNull(message="AMENITY_ID must not be null.")
-    private Long amenityId;
-    @NotNull(message="AMENITY name must not be null.")
-    private String amenity;
-
-    public AmenityDTO(Long amenityId, String amenity) {
-        this.amenityId = amenityId;
-        this.amenity = amenity;
-    }
-
-    public Long getAmenityId() {
-        return amenityId;
-    }
-
-    public void setAmenityId(Long amenityId) {
-        this.amenityId = amenityId;
-    }
-
-    public String getAmenity() {
-        return amenity;
-    }
-
-    public void setAmenity(String amenity) {
-        this.amenity = amenity;
-    }
-
-    public static AmenityDTO prepareAmenityDTO(Amenity amenity) {
-        return new AmenityDTO(amenity.getAmenityId(), amenity.getAmenity());
-    }
-
-    public static Amenity prepareAmenityEntity(AmenityDTO amenityDTO) {
-        Amenity amenity = new Amenity();
-        amenity.setAmenityId(amenityDTO.getAmenityId());
-        amenity.setAmenity(amenityDTO.getAmenity());
-        return amenity;
-    }
-}
