--- src/main/java/com/dart/explore/entity/StationColor.java
@@ -1,5 +1,0 @@
-package com.dart.explore.entity;
-
-public enum StationColor {
-    BLUE, RED, ORANGE, GREEN
-}
