--- src/main/java/com/dart/explore/exception/DartExploreException.java
@@ -1,7 +1,0 @@
-package com.dart.explore.exception;
-
-public class DartExploreException extends Exception {
-    public DartExploreException(String message){
-        super(message);
-    }
-}
