--- src/main/java/com/dart/explore/repository/AmenityRepository.java
@@ -1,15 +1,0 @@
-package com.dart.explore.repository;
-
-import com.dart.explore.entity.Amenity;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
-
-import java.util.List;
-
-public interface AmenityRepository extends CrudRepository<Amenity, Long> {
-    @Query("SELECT a FROM Amenity a")
-    List<Amenity> findAllAmenities();
-    @Query("SELECT a FROM Amenity a WHERE a.amenityId IN :ids")
-    List<Amenity> findAllAmenitiesById(@Param("ids") List<Long> ids);
-}
